add LimbPanel
