+++ src/main/java/com/marginallyclever/ro3/node/nodes/pose/Limb.java
@@ -47,8 +47,8 @@ public class Limb extends Pose {
     /**
      * @return the end effector pose or null if not set.
      */
-    public Pose getEndEffector() {
-        return endEffector.getSubject() == null ? null : endEffector.getSubject();
+    public NodePath<Pose> getEndEffector() {
+        return endEffector;
     }
 
     public List<NodePath<Motor>> getMotors() {
@@ -63,8 +63,14 @@ public class Limb extends Pose {
         return count;
     }
 
-    public Motor getJoint(int i) {
-        return motors.get(i).getSubject();
+    /**
+     * Get the motor at the given index.
+     * @param index the index of the motor to get.
+     * @return the motor at the given index.
+     * @throws IndexOutOfBoundsException if the index is out of range.
+     */
+    public Motor getJoint(int index) {
+        return motors.get(index).getSubject();
     }
 
     public double[] getAllJointAngles() {
@@ -77,6 +83,16 @@ public class Limb extends Pose {
             }
         }
         return result;
+    }
+
+    /**
+     * Get the motor at the given index.
+     * @param index the index of the motor to get.
+     * @param newValue the new motor to set.
+     * @throws IndexOutOfBoundsException if the index is out of range.
+     */
+    public void setJoint(int index, Motor newValue) {
+        motors.get(index).setRelativePath(this,newValue);
     }
 
     public void setAllJointAngles(double[] values) {
@@ -160,96 +176,7 @@ public class Limb extends Pose {
 
     @Override
     public void getComponents(List<JPanel> list) {
-        JPanel pane = new JPanel(new GridBagLayout());
-        list.add(pane);
-        pane.setName(Limb.class.getSimpleName());
-
-        GridBagConstraints gbc = new GridBagConstraints();
-        gbc.weightx = 1.0;
-        gbc.weighty = 1.0;
-        gbc.fill = GridBagConstraints.BOTH;
-        gbc.gridx=0;
-        gbc.gridy=0;
-        gbc.gridwidth=1;
-        NodePanelHelper.addNodeSelector(pane, "End Effector", endEffector, Pose.class, gbc,this);
-
-        gbc.gridx=0;
-        gbc.gridwidth=2;
-        gbc.gridy++;
-        pane.add(createFKDials(),gbc);
-
-        gbc.gridy++;
-        pane.add(addMotorPanel(),gbc);
-
+        list.add(new LimbPanel(this));
         super.getComponents(list);
-    }
-
-    private JComponent createFKDials() {
-        var containerPanel = new CollapsiblePanel("Forward Kinematics");
-        var outerPanel = containerPanel.getContentPane();
-        outerPanel.setLayout(new GridLayout(0,3));
-
-        int count=0;
-        for(int i=0;i<getNumJoints();++i) {
-            Motor motor = motors.get(i).getSubject();
-            if(motor==null) continue;
-            outerPanel.add(createOneFKDial(motor));
-            count++;
-        }
-        count = 3-(count%3);
-        for(int i=0;i<count;++i) {
-            outerPanel.add(new JPanel());
-        }
-
-        return containerPanel;
-    }
-
-    private JPanel createOneFKDial(final Motor motor) {
-        JPanel panel = new JPanel(new BorderLayout());
-        Dial dial = new Dial();
-        dial.addActionListener(e -> {
-            if(motor.hasHinge()) {
-                motor.getHinge().setAngle(dial.getValue());
-                dial.setValue(motor.getHinge().getAngle());
-            }
-        });
-        // TODO subscribe to motor.getAxle().getAngle(), then dial.setValue() without triggering an action event.
-
-        JLabel label = new JLabel(motor.getName());
-        label.setLabelFor(dial);
-        label.setHorizontalAlignment(SwingConstants.CENTER);
-        panel.add(label,BorderLayout.PAGE_START);
-        panel.add(dial,BorderLayout.CENTER);
-        dial.setPreferredSize(new Dimension(80,80));
-        if(motor.hasHinge()) {
-            dial.setValue(motor.getHinge().getAngle());
-        }
-        return panel;
-    }
-
-    private JComponent addMotorPanel() {
-        var containerPanel = new CollapsiblePanel("Motors");
-        containerPanel.setCollapsed(true);
-        var outerPanel = containerPanel.getContentPane();
-        outerPanel.setLayout(new GridBagLayout());
-
-        GridBagConstraints gbc = new GridBagConstraints();
-        gbc.weightx = 1.0;
-        gbc.weighty = 1.0;
-        gbc.gridx=0;
-        gbc.gridy=0;
-        gbc.fill = GridBagConstraints.BOTH;
-
-        // add a selector for each motor
-        var motorSelector = new NodeSelector[MAX_JOINTS];
-        for(int i=0;i<MAX_JOINTS;++i) {
-            motorSelector[i] = new NodeSelector<>(Motor.class, motors.get(i).getSubject());
-            int j = i;
-            motorSelector[i].addPropertyChangeListener("subject",(e)-> {
-                motors.get(j).setRelativePath(this,(Motor)e.getNewValue());
-            });
-            NodePanelHelper.addLabelAndComponent(outerPanel, "Motor "+i, motorSelector[i],gbc);
-        }
-        return containerPanel;
     }
 }
